+++ logzio-sender/src/main/java/io/logz/sender/LogzioSender.java
@@ -5,10 +5,7 @@ import com.google.gson.JsonObject;
 import io.logz.sender.exceptions.LogzioParameterErrorException;
 import io.logz.sender.exceptions.LogzioServerErrorException;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
-import java.io.OutputStream;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -21,7 +18,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.zip.GZIPOutputStream;
 
 public class LogzioSender {
     private static final int MAX_SIZE_IN_BYTES = 3 * 1024 * 1024;  // 3 MB
@@ -39,7 +35,6 @@ public class LogzioSender {
     private ScheduledExecutorService tasksExecutor;
     private final int gcPersistedQueueFilesIntervalSeconds;
     private final AtomicBoolean drainRunning = new AtomicBoolean(false);
-    private final HttpsRequestConfiguration httpsRequestConfiguration;
     private final HttpsSyncSender httpsSyncSender;
 
     private LogzioSender(String logzioToken, String logzioType, int drainTimeout, int fsPercentThreshold, File bufferDir,
@@ -48,7 +43,7 @@ public class LogzioSender {
                          int gcPersistedQueueFilesIntervalSeconds, boolean compressRequests)
             throws  LogzioParameterErrorException {
 
-        httpsRequestConfiguration = HttpsRequestConfiguration
+        HttpsRequestConfiguration httpsRequestConfiguration = HttpsRequestConfiguration
                 .builder()
                 .setLogzioToken(logzioToken)
                 .setLogzioType(logzioType)
@@ -198,11 +193,13 @@ public class LogzioSender {
 
     private List<FormattedLogMessage> dequeueUpToMaxBatchSize() {
         List<FormattedLogMessage> logsList = new ArrayList<>();
+        int totalSize = 0;
         while (!logsBuffer.isEmpty()) {
             byte[] message  = logsBuffer.dequeue();
             if (message != null && message.length > 0) {
                 logsList.add(new FormattedLogMessage(message));
-                if (sizeInBytes(logsList) >= MAX_SIZE_IN_BYTES) {
+                totalSize += message.length;
+                if (totalSize >= MAX_SIZE_IN_BYTES) {
                     break;
                 }
             }
@@ -216,7 +213,7 @@ public class LogzioSender {
             while (!logsBuffer.isEmpty()) {
                 List<FormattedLogMessage> logsList = dequeueUpToMaxBatchSize();
                 try {
-                    httpsSyncSender.sendToLogzio(toNewLineSeparatedByteArray(logsList));
+                    httpsSyncSender.sendToLogzio(logsList);
 
                 } catch (LogzioServerErrorException e) {
                     debug("Could not send log to logz.io: ", e);
@@ -233,25 +230,6 @@ public class LogzioSender {
                     break;
                 }
             }
-        }
-    }
-
-    private int sizeInBytes(List<FormattedLogMessage> logMessages) {
-        int totalSize = 0;
-        for (FormattedLogMessage currLog : logMessages) totalSize += currLog.getSize();
-
-        return totalSize;
-    }
-
-    private byte[] toNewLineSeparatedByteArray(List<FormattedLogMessage> messages) {
-        try (ByteArrayOutputStream byteOutputStream = new ByteArrayOutputStream(sizeInBytes(messages));
-             OutputStream os = httpsRequestConfiguration.isCompressRequests() ? new GZIPOutputStream(byteOutputStream) : byteOutputStream) {
-            for (FormattedLogMessage currMessage : messages) os.write(currMessage.getMessage());
-            // Need close before return for gzip compression, The stream only knows to compress and write the last bytes when you tell it to close
-            os.close();
-            return byteOutputStream.toByteArray();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
